embedded/requirements.txt
beautifulsoup4==4.12.3
bidict==0.22.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
docopt==0.6.2
dotenv==0.0.5
fernet==1.0.1
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-speech==2.24.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
idna==3.6
Jinja2==3.1.3
Js2Py==0.74
MarkupSafe==2.1.5
numpy==1.26.3
packaging==23.2
pipwin==0.5.2
proto-plus==1.23.0
protobuf==4.25.2
pvporcupine==3.0.2
pvrecorder==1.2.2
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pygame==2.5.2
pyjsparser==2.7.1
PyPrind==2.11.3
pyproject==1.3.1
pyserial==3.5
pySmartDL==1.3.4
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
requests==2.31.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
tzlocal==5.2
urllib3==2.2.0
Wave==0.0.2
websocket-client==1.7.0
wsproto==1.2.0
